+++ sources/misc/ft_strlen.c
@@ -6,13 +6,15 @@
 /*   By: emedea <emedea@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/04/04 21:33:41 by emedea            #+#    #+#             */
-/*   Updated: 2019/08/12 15:57:30 by emedea           ###   ########.fr       */
+/*   Updated: 2019/05/17 10:36:32 by emedea           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-int		ft_strlen(const char *s)
+#include <string.h>
+
+size_t	ft_strlen(const char *s)
 {
-	int	i;
+	size_t i;
 
 	i = 0;
 	while (s[i])
